requirements.txt
<<<<<<< HEAD
altgraph==0.17.3
asttokens==2.0.8
backcall==0.2.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
six==1.16.0
stack-data==0.5.0
tornado==6.2
traitlets==5.4.0
wcwidth==0.2.5
=======
Pillow==9.2.0
>>>>>>> e8b3520afa44ff27cd38e5518a508a1c7cb43b4f
